Address review feedback.

+++ qrenderdoc/3rdparty/scintilla/src/PositionCache.cxx
@@ -391,7 +391,7 @@ static inline int KeyFromString(const char *charBytes, size_t len) {
 }
 
 SpecialRepresentations::SpecialRepresentations() {
-	std::fill(startByteHasReprs, startByteHasReprs+0x100, (short)0);
+	std::fill(startByteHasReprs, startByteHasReprs+0x100, 0);
 }
 
 void SpecialRepresentations::SetRepresentation(const char *charBytes, const char *value) {
@@ -432,7 +432,7 @@ bool SpecialRepresentations::Contains(const char *charBytes, size_t len) const {
 
 void SpecialRepresentations::Clear() {
 	mapReprs.clear();
-	std::fill(startByteHasReprs, startByteHasReprs+0x100, (short)0);
+	std::fill(startByteHasReprs, startByteHasReprs+0x100, 0);
 }
 
 void BreakFinder::Insert(int val) {
